Factors out URL fetcher into its own hook
